add credentials and get discussion of tweet

+++ Praktisch/tweetConversations/src/main/java/com/tweetConversations/tweetConversations/model/Tweet.java
@@ -17,7 +17,7 @@ public class Tweet {
 
 	private String title;
 
-	private Integer tweetId;
+	private long tweetId;
 
 	@Relationship(type="REPLYTO", direction=Relationship.Direction.OUTGOING)
 	private Tweet reply;
